Split params into separate files
